refactor(process): extract loadEnvFile to process_env_file.rs (under 2000-cap after merge)
