+++ crates/perry-runtime/src/lib.rs
@@ -63,6 +63,7 @@ pub mod node_submodules;
 pub mod node_test;
 // #3137/#3138/#3142: public `node:v8` serialize/deserialize + heap stats + GCProfiler.
 pub mod node_v8;
+mod process_env_file;
 // #2935: surface the zlib option-level resolver at the crate root so
 // perry-stdlib's bundled codecs (and the `perry-ext-zlib` extern) can reach it.
 pub use node_submodules::js_zlib_resolve_level;
